requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
